Xonsh: fix mkvenv/activate

+++ .config/xonsh/aliases.py
@@ -20,33 +20,33 @@ def _mkvenv():
     """
     Create a virtual env, activate source it, and pip install requirements.txt.
     """
+    from os.path import exists
+    # Create a vox virtual environment and activate it
     vox new .venv
     vox activate .venv
+    # Upgrade pip
     pip install --upgrade pip
-    pip install -r requirements.txt
+    # Install requirememts if available
+    if exists("requirements.txt"):
+        pip install -r requirements.txt
+
 
 def _activate():
     """
-    Search recursively for a vox .venv, and if found activate it.
+    Search in parent directories for a .venv, and if found activate it with vox.
     """
+    from pathlib import Path
+    curr_path = Path.cwd()
+    #TODO this actually skips the root directory
+    while curr_path.parent != curr_path:
+        venv_path = curr_path.joinpath(".venv")
+        if venv_path.is_dir():
+            vox activate @(str(venv_path))
+            return 0
+        curr_path = curr_path.parent
+    print("Could not find a .venv to activate.")
+    return 1
 
 aliases['mkvenv'] = _mkvenv
 aliases["deactivate"] = "vox deactivate"
 aliases["activate"] = _activate
-
-"""
-activate() {
-        local dir=$PWD
-        while [ "$dir" != "/" ]; do
-                if [ -f "$dir/.venv/bin/activate" ]; then
-                        break
-                fi
-                dir=$(dirname $dir)
-        done
-        if [ -f "$dir/.venv/bin/activate" ]; then
-                source "$dir/.venv/bin/activate"
-        else
-                "Could not find .venv"
-        fi
-}
-"""
